Rename get_font_format and move it to __init__.py
To be reused.
