+++ python/psautohint/__main__.py
@@ -8,13 +8,13 @@ from __future__ import print_function, absolute_import
 
 import argparse
 import os
-import plistlib
 import re
 import sys
 import textwrap
 
 from fontTools.misc.py23 import open
 
+from psautohint import get_font_format
 from psautohint._psautohint import version as PSAUTOHINT_VERSION
 from psautohint.autohint import (ACFontError, ACHintError, ACOptions,
                                  hintFiles, logMsg)
@@ -416,38 +416,6 @@ def _check_save_path(path_str):
     return test_path
 
 
-def _font_is_ufo(path):
-    if os.path.isdir(path) and path.lower().endswith('.ufo'):
-        meta_path = os.path.join(path, 'metainfo.plist')
-        if os.path.isfile(meta_path):
-            metainfo = plistlib.readPlist(meta_path)
-            if all([key in metainfo for key in ('creator', 'formatVersion')]):
-                return True
-    return False
-
-
-def _get_font_format(font_file_path):
-    if os.path.isfile(font_file_path):
-        with open(font_file_path, 'rb') as f:
-            head = f.read(4)
-            if head == b'OTTO':
-                return 'OTF'
-            elif head[0:2] == b'\x01\x00':
-                return 'CFF'
-            elif head == b'\x80\x01\xb6\x12':
-                return 'PFB'
-            elif head in (b'%!PS', b'%!Fo'):
-                for fullhead in (b'%!PS-AdobeFont', b'%!FontType1'):
-                    f.seek(0)
-                    if f.read(len(fullhead)) == fullhead:
-                        return 'PFA'
-        return None
-    else:
-        if _font_is_ufo(font_file_path):
-            return "UFO"
-        return None
-
-
 def _validate_font_paths(path_lst, parser):
     """
     Checks that all input paths are fonts, and that all are the same format.
@@ -455,7 +423,7 @@ def _validate_font_paths(path_lst, parser):
     """
     format_set = set()
     for path in path_lst:
-        font_format = _get_font_format(path)
+        font_format = get_font_format(path)
         if font_format is None:
             parser.error("{} is not a supported font format".format(
                 os.path.basename(path)))
